+++ android/src/main/kotlin/top/kikt/imagescanner/core/entity/AssetEntity.kt
@@ -10,6 +10,7 @@ data class AssetEntity(
   val createDt: Long,
   val width: Int,
   val height: Int,
+  //
   val type: Int,
   val displayName: String,
   val modifiedDate: Long,